add missing files

+++ src/game.h
@@ -4,7 +4,7 @@
 #include "types.h"
 
 void game(tinyint player_count);
-bool triggerNewTurn(int player_index, GameState* state);
+bool triggerNewTurn(tinyint player_index, GameState* state);
 void scan(Player* player, Player* target);
 void asktargetedcell(Player* player, tinyint* x, tinyint* y);
